updated rotation model to be based on relative position and not drift

+++ src/tabcontrollers/MoveCenterTabController.h
@@ -171,7 +171,7 @@ private:
     vr::HmdQuaternion_t m_hmdQuaternion;
     double m_hmdYawTotal = 0.0;
     double m_hmdYawSingle = 0.0;
-    double m_hmdYawTurnCount = 0;
+    int m_hmdYawTurnCount = 0;
     vr::ETrackedControllerRole m_activeDragHand
         = vr::TrackedControllerRole_Invalid;
     vr::ETrackedControllerRole m_lastMoveHand
